requirements.txt
Django==5.2.3
django-allauth==0.61.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0